Add converters.
Update the internal architecture to be more flexible by adding the concept of converters. It is now possible to register a custom converter either globally or on the current instance of PlayDataTables to be able to handle a type natively.

The previously handled types are now built-in converters and are directly registered on the the global configuration.

Fixes #2

Signed-off-by: Pierre Adam <pierre.adam@payintech.com>